+++ lib/roble.dart
@@ -5,3 +5,4 @@ export 'src/roble_api_config.dart';
 export 'src/roble_api_database_impl.dart';
 export 'src/roble_api_exception.dart';
 export 'src/roble_models.dart';
+export 'src/roble_storage.dart';